% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_clambda}
\alias{c_clambda}
\title{Tuning lambda assuming GIC to be a constant function in Rcpp}
\usage{
c_clambda(y, z, mu, learning_rate = 0.01, max_iter = 1000L, tol = 1e-06)
}
\arguments{
\item{y}{Response variable which is a n*1 vector}

\item{z}{Designed n*p matrix of composition as explanatory variables}

\item{mu}{Penalty parameter avoiding zero-sum constraint from affecting convergence(suggested to be 1)}

\item{learning_rate}{Learning rate for upgrading grad function}

\item{max_iter}{Iteration stop condition: Max iterations}

\item{tol}{Iteration stop condition: Tolerance}
}
\description{
Specific information can be found in 'r_clambda' function(It has the same problem as 'c_cmdlasso' which means you need to make sure that y and z is numeric before using)
}
\examples{
\dontrun{
data(data)
y <- as.vector(unlist(data[1]))
z <- as.matrix(data[2:31])
mu <- 1
c_clambda(y,z,mu)
}
}
