% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{r_cdmlasso}
\alias{r_cdmlasso}
\title{A Coordinate Descend Algorithm for Compositional Data Lasso problem in R}
\usage{
r_cdmlasso(y, z, lambda, mu, tol = 1e-05, max_iter = 1000)
}
\arguments{
\item{y}{Response variable which is a n*1 vector}

\item{z}{Designed n*p matrix of composition as explanatory variables}

\item{lambda}{Regularization parameter in Lasso problem}

\item{mu}{Penalty parameter avoiding zero-sum constraint from affecting convergence(suggested to be 1)}

\item{tol}{Iteration stop condition: Tolerance}

\item{max_iter}{Iteration stop condition: Max iterations}
}
\value{
A p*1 vector of regression coefficient estimator
}
\description{
This function proposed a coordinate descend algorithm in high dimensional lasso regression problems.
}
\examples{
\dontrun{
data(data)
y <- data[1]
z <- data[2:31]
mu <- 1; lambda <- 0.5
r_cdmlasso(y,z,lambda,mu)
}
}
