% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.tuning.R
\name{r_clambda}
\alias{r_clambda}
\title{Tuning lambda assuming GIC to be a constant function in R}
\usage{
r_clambda(y, z, mu)
}
\arguments{
\item{y}{Response variable which is a n*1 vector}

\item{z}{Designed n*p matrix of composition as explanatory variables}

\item{mu}{Penalty parameter avoiding zero-sum constraint from affecting convergence}
}
\value{
A numerical solve making GIC to reach its minimum
}
\description{
Under the GIC standard, find the best lambda for the lasso regression
}
\examples{
\dontrun{
data(data)
y <- data[1]
z <- data[2:31]
mu <- 1
r_clambda(y,z,mu)
}
}
