% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.tuning.R
\name{r_slambda}
\alias{r_slambda}
\title{Tuning lambda using a prepared series within GIC standard in R}
\usage{
r_slambda(y, z, mu, low = 0, up = 5, by = 0.1)
}
\arguments{
\item{y}{Response variable which is a n*1 vector}

\item{z}{Designed n*p matrix of composition as explanatory variables}

\item{mu}{Penalty parameter avoiding zero-sum constraint from affecting convergence}

\item{low}{Lower bound of the interval of prepared lambda}

\item{up}{Upper bound of the interval of prepared lambda}

\item{by}{Step size of the interval of prepared lambda}
}
\value{
The best point within the series to minimum GIC
}
\description{
Under the GIC standard, find the best lambda for the lasso regression
}
\examples{
\dontrun{
data(data)
y <- data[1]
z <- data[2:31]
mu <- 1
r_slambda(y,z,mu,low = 0,up = 2)
}
}
