\name{SA24204132-package}
\alias{SA24204132-package}
\alias{SA24204132}
\docType{package}
\title{
  A Lasso regression package focusing on compositonal data.
}
\description{
  Several R and Rcpp functions are included in this package for regression and
  tuning parameter.
}
\author{
Zehua Dong

Maintainer: Zehua Dong <1758203654@qq.com>
}
\keyword{ package }