% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softthreshold.R
\name{soft_thr}
\alias{soft_thr}
\title{Soft threshold function}
\usage{
soft_thr(t, lambda)
}
\arguments{
\item{t}{A specific form in the operation term}

\item{lambda}{Regularization parameter in Lasso problem}
}
\value{
The Lasso Soft-threshold operation term computed by samples
}
\description{
Soft threshold function
}
